requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==38.0.3
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pkg_resources==0.0.0
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.6
sqlparse==0.4.3
